lab1/inference/requirements.txt
numpy
boto3
torch
Pillow